requirements.txt
streamlit
plotly
fpdf2
kaleido
pandas
numpy
scikit-learn
torch
